Object Interact More Customization
I made it so now you can choose what side of the object you want the player to interact with. I wrote the correct inputs for each side of the object in the comments of ObjectInteract. If this code broke anything please let me know and I can revert it.